IService_Server/requirements.txt
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==3.2
djangorestframework
markdown
django-filter
django-phonenumber-field
djangorestframework-jwt
